validate raw data
