+++ lib/dataContract/DataContract.js
@@ -11,9 +11,10 @@ class DataContract {
    * @param {RawDataContract} rawDataContract
    */
   constructor(rawDataContract) {
+    this.protocolVersion = rawDataContract.protocolVersion;
+
     this.id = EncodedBuffer.from(rawDataContract.$id, EncodedBuffer.ENCODING.BASE58);
     this.ownerId = EncodedBuffer.from(rawDataContract.ownerId, EncodedBuffer.ENCODING.BASE58);
-    this.protocolVersion = rawDataContract.protocolVersion;
 
     this.setJsonMetaSchema(rawDataContract.$schema);
     this.setDocuments(rawDataContract.documents);
@@ -184,6 +185,7 @@ class DataContract {
   }
 
   /**
+   * Return Data Contract as plain object
    *
    * @return {RawDataContract}
    */
@@ -206,9 +208,9 @@ class DataContract {
   }
 
   /**
-   * Return Data Contract as plain object
+   * Return Data Contract as JSON object
    *
-   * @return {Object}
+   * @return {JsonDataContract}
    */
   toJSON() {
     const json = {
@@ -229,12 +231,12 @@ class DataContract {
   }
 
   /**
-   * Return serialized Data Contract
+   * Return Data Contract as a Buffer
    *
-   * @return {Buffer}
+   * @returns {Buffer}
    */
-  serialize() {
-    return encode(this.toJSON());
+  toBuffer() {
+    return encode(this.toObject());
   }
 
   /**
@@ -278,10 +280,29 @@ class DataContract {
  * @property {Object<string, Object>} [definitions]
  */
 
+/**
+ * @typedef {Object} JsonDataContract
+ * @property {number} protocolVersion
+ * @property {string} $id
+ * @property {string} $schema
+ * @property {string} ownerId
+ * @property {Object<string, Object>} documents
+ * @property {Object<string, Object>} [definitions]
+ */
+
 DataContract.PROTOCOL_VERSION = 0;
 
 DataContract.DEFAULTS = {
   SCHEMA: 'https://schema.dash.org/dpp-0-4-0/meta/data-contract',
+};
+
+DataContract.ENCODED_PROPERTIES = {
+  $id: {
+    contentEncoding: 'base58',
+  },
+  ownerId: {
+    contentEncoding: 'base58',
+  },
 };
 
 module.exports = DataContract;
