+++ lib/dataContract/DataContractFacade.js
@@ -75,7 +75,7 @@ class DataContractFacade {
    * @return {Promise<ValidationResult>}
    */
   async validate(dataContract) {
-    return this.validateDataContract(dataContract);
+    return this.validateDataContract(dataContract.toObject());
   }
 }
