+++ lib/dataContract/DataContractFactory.js
@@ -101,7 +101,7 @@ class DataContractFactory {
     return new DataContractCreateTransition({
       protocolVersion: DataContract.PROTOCOL_VERSION,
       dataContract: dataContract.toObject(),
-      entropy: dataContract.getEntropy().toBuffer(),
+      entropy: dataContract.getEntropy(),
     });
   }
 }
