+++ lib/dataContract/enrichDataContractWithBaseSchema.js
@@ -18,9 +18,7 @@ function enrichDataContractWithBaseSchema(
   schemaIdPrefix = '',
   excludeProperties = [],
 ) {
-  const rawDataContract = dataContract.toJSON();
-
-  const clonedDataContract = lodashCloneDeep(rawDataContract);
+  const clonedDataContract = lodashCloneDeep(dataContract.toObject());
 
   delete clonedDataContract.$schema;
 
@@ -54,7 +52,10 @@ function enrichDataContractWithBaseSchema(
   });
 
   // Add schema $id prefix
-  clonedDataContract.$id = schemaIdPrefix + clonedDataContract.$id;
+  clonedDataContract.$id = Buffer.concat([
+    Buffer.from(schemaIdPrefix),
+    clonedDataContract.$id,
+  ]);
 
   return new DataContract(clonedDataContract);
 }
