+++ lib/dataContract/stateTransition/DataContractCreateTransition.js
@@ -65,7 +65,7 @@ class DataContractCreateTransition extends AbstractStateTransitionIdentitySigned
    * @param {Object} [options]
    * @param {boolean} [options.skipSignature]
    * @param {boolean} [options.encodedBuffer=false]
-   * @return {Object}
+   * @return {RawDataContractCreateTransition}
    */
   toObject(options = {}) {
     Object.assign(
@@ -76,20 +76,17 @@ class DataContractCreateTransition extends AbstractStateTransitionIdentitySigned
       },
     );
 
-    let plainObject = {
+    const rawStateTransition = {
       ...super.toObject(options),
       dataContract: this.getDataContract().toObject(),
-      entropy: this.entropy,
+      entropy: this.getEntropy(),
     };
 
     if (!options.encodedBuffer) {
-      plainObject = {
-        ...plainObject,
-        entropy: plainObject.entropy.toBuffer(),
-      };
+      rawStateTransition.entropy = this.getEntropy().toBuffer();
     }
 
-    return plainObject;
+    return rawStateTransition;
   }
 
   /**
@@ -97,14 +94,14 @@ class DataContractCreateTransition extends AbstractStateTransitionIdentitySigned
    *
    * @param {Object} [options]
    *
-   * @return {Object}
+   * @return {JsonDataContractCreateTransition}
    */
   toJSON(options = {}) {
-    const json = super.toJSON(options);
+    const jsonStateTransition = super.toJSON(options);
 
     return {
-      ...json,
-      entropy: json.entropy.toString(),
+      ...jsonStateTransition,
+      entropy: jsonStateTransition.entropy.toString(),
       dataContract: this.getDataContract().toJSON(),
     };
   }
@@ -138,7 +135,24 @@ class DataContractCreateTransition extends AbstractStateTransitionIdentitySigned
  * @property {RawDataContract} dataContract
  * @property {number|null} signaturePublicKeyId
  * @property {Buffer|null} signature
- * @property {string|null} entropy
+ * @property {Buffer} entropy
  */
+
+/**
+ * @typedef {Object} JsonDataContractCreateTransition
+ * @property {number} protocolVersion
+ * @property {number} type
+ * @property {JsonDataContract} dataContract
+ * @property {number|null} signaturePublicKeyId
+ * @property {string|null} signature
+ * @property {string} entropy
+ */
+
+DataContractCreateTransition.ENCODED_PROPERTIES = {
+  ...AbstractStateTransitionIdentitySigned.ENCODED_PROPERTIES,
+  entropy: {
+    contentEncoding: 'base58',
+  },
+};
 
 module.exports = DataContractCreateTransition;
