+++ lib/dataContract/validateDataContractFactory.js
@@ -14,6 +14,7 @@ const InvalidIndexedPropertyConstraintError = require('../errors/InvalidIndexedP
 const InvalidCompoundIndexError = require('../errors/InvalidCompoundIndexError');
 
 const getPropertyDefinitionByPath = require('./getPropertyDefinitionByPath');
+const encodeObjectProperties = require('../util/encoding/encodeObjectProperties');
 
 const allowedSystemProperties = ['$id', '$ownerId', '$createdAt', '$updatedAt'];
 const prebuiltIndices = ['$id'];
@@ -33,18 +34,16 @@ module.exports = function validateDataContractFactory(
 ) {
   /**
    * @typedef validateDataContract
-   * @param {DataContract|RawDataContract} dataContract
+   * @param {RawDataContract} rawDataContract
    * @return {ValidationResult}
    */
-  async function validateDataContract(dataContract) {
-    if (!(dataContract instanceof DataContract)) {
-      // eslint-disable-next-line no-param-reassign
-      dataContract = new DataContract(dataContract);
-    }
-
-    const jsonDataContract = dataContract.toJSON();
-
+  async function validateDataContract(rawDataContract) {
     const result = new ValidationResult();
+
+    const jsonDataContract = encodeObjectProperties(
+      rawDataContract,
+      DataContract.ENCODED_PROPERTIES,
+    );
 
     // Validate Data Contract schema
     result.merge(
@@ -59,12 +58,12 @@ module.exports = function validateDataContractFactory(
     }
 
     result.merge(
-      await validateDataContractMaxDepth(jsonDataContract),
+      await validateDataContractMaxDepth(rawDataContract),
     );
 
     // Validate Document JSON Schemas
     const enrichedDataContract = enrichDataContractWithBaseSchema(
-      dataContract,
+      new DataContract(rawDataContract),
       baseDocumentSchema,
       'documentBase',
     );
@@ -111,7 +110,7 @@ module.exports = function validateDataContractFactory(
               isUniqueIndexLimitReached = true;
 
               result.addError(new UniqueIndicesLimitReachedError(
-                jsonDataContract,
+                rawDataContract,
                 documentType,
               ));
             }
@@ -125,7 +124,7 @@ module.exports = function validateDataContractFactory(
 
               if (isSingleIndex) {
                 result.addError(new SystemPropertyIndexAlreadyPresentError(
-                  jsonDataContract,
+                  rawDataContract,
                   documentType,
                   indexDefinition,
                   propertyName,
@@ -150,7 +149,7 @@ module.exports = function validateDataContractFactory(
             .map(([propertyName]) => {
               result.addError(
                 new UndefinedIndexPropertyError(
-                  jsonDataContract,
+                  rawDataContract,
                   documentType,
                   indexDefinition,
                   propertyName,
@@ -185,7 +184,7 @@ module.exports = function validateDataContractFactory(
 
             if (invalidPropertyType) {
               result.addError(new InvalidIndexPropertyTypeError(
-                jsonDataContract,
+                rawDataContract,
                 documentType,
                 indexDefinition,
                 propertyName,
@@ -199,7 +198,7 @@ module.exports = function validateDataContractFactory(
               if (maxLength === undefined) {
                 result.addError(
                   new InvalidIndexedPropertyConstraintError(
-                    jsonDataContract,
+                    rawDataContract,
                     documentType,
                     indexDefinition,
                     propertyName,
@@ -212,7 +211,7 @@ module.exports = function validateDataContractFactory(
               if (maxLength !== undefined && maxLength > MAX_INDEXED_STRING_PROPERTY_LENGTH) {
                 result.addError(
                   new InvalidIndexedPropertyConstraintError(
-                    jsonDataContract,
+                    rawDataContract,
                     documentType,
                     indexDefinition,
                     propertyName,
@@ -245,7 +244,7 @@ module.exports = function validateDataContractFactory(
           if (indicesFingerprints.includes(indicesFingerprint)) {
             result.addError(
               new DuplicateIndexError(
-                jsonDataContract,
+                rawDataContract,
                 documentType,
                 indexDefinition,
               ),
