+++ lib/dataTrigger/dpnsTriggers/createDomainDataTrigger.js
@@ -54,7 +54,7 @@ async function createDomainDataTrigger(documentTransition, context, topLevelIden
   }
 
   if (records.dashUniqueIdentityId
-    && Buffer.compare(context.getOwnerId(), records.dashUniqueIdentityId) !== 0
+    && !context.getOwnerId().equals(records.dashUniqueIdentityId)
   ) {
     result.addError(
       new DataTriggerConditionError(
@@ -67,7 +67,7 @@ async function createDomainDataTrigger(documentTransition, context, topLevelIden
   }
 
   if (records.dashAliasIdentityId
-    && Buffer.compare(context.getOwnerId(), records.dashAliasIdentityId) !== 0) {
+    && !context.getOwnerId().equals(records.dashAliasIdentityId)) {
     result.addError(
       new DataTriggerConditionError(
         documentTransition,
@@ -79,7 +79,7 @@ async function createDomainDataTrigger(documentTransition, context, topLevelIden
   }
 
   if (normalizedParentDomainName.length === 0
-    && Buffer.compare(context.getOwnerId().toBuffer(), topLevelIdentity) !== 0) {
+    && !context.getOwnerId().equals(topLevelIdentity)) {
     result.addError(
       new DataTriggerConditionError(
         documentTransition,
@@ -132,7 +132,7 @@ async function createDomainDataTrigger(documentTransition, context, topLevelIden
     }
 
     if (parentDomain.getData().subdomainRules.allowSubdomains === false
-      && Buffer.compare(context.getOwnerId(), parentDomain.getOwnerId()) !== 0) {
+      && !context.getOwnerId().equals(parentDomain.getOwnerId())) {
       result.addError(
         new DataTriggerConditionError(
           documentTransition,
