+++ lib/document/DocumentFactory.js
@@ -217,7 +217,7 @@ class DocumentFactory {
 
     const mismatchedOwnerIdsLength = documentsFlattened
       .reduce((result, document) => {
-        if (Buffer.compare(document.getOwnerId(), ownerId) !== 0) {
+        if (!document.getOwnerId().equals(ownerId)) {
           // eslint-disable-next-line no-param-reassign
           result += 1;
         }
