+++ lib/document/stateTransition/DocumentsBatchTransition.js
@@ -168,4 +168,8 @@ class DocumentsBatchTransition extends AbstractStateTransitionIdentitySigned {
  * @property {Buffer|null} signature
  */
 
+DocumentsBatchTransition.ENCODED_PROPERTIES = {
+  ...AbstractStateTransitionIdentitySigned.ENCODED_PROPERTIES,
+};
+
 module.exports = DocumentsBatchTransition;
