+++ lib/document/stateTransition/documentTransition/AbstractDocumentTransition.js
@@ -131,4 +131,10 @@ AbstractDocumentTransition.ACTION_NAMES = {
   DELETE: 'delete',
 };
 
+AbstractDocumentTransition.ENCODED_PROPERTIES = {
+  dataContractId: {
+    contentEncoding: 'base58',
+  },
+};
+
 module.exports = AbstractDocumentTransition;
