+++ lib/document/stateTransition/documentTransition/DocumentCreateTransition.js
@@ -205,4 +205,14 @@ class DocumentCreateTransition extends AbstractDocumentTransition {
 
 DocumentCreateTransition.INITIAL_REVISION = 1;
 
+DocumentCreateTransition.ENCODED_PROPERTIES = {
+  ...AbstractDocumentTransition.ENCODED_PROPERTIES,
+  $id: {
+    contentEncoding: 'base58',
+  },
+  $entropy: {
+    contentEncoding: 'base58',
+  },
+};
+
 module.exports = DocumentCreateTransition;
