+++ lib/document/stateTransition/documentTransition/DocumentDeleteTransition.js
@@ -96,4 +96,11 @@ class DocumentDeleteTransition extends AbstractDocumentTransition {
  * @property {string} $type
  */
 
+DocumentDeleteTransition.ENCODED_PROPERTIES = {
+  ...AbstractDocumentTransition.ENCODED_PROPERTIES,
+  $id: {
+    contentEncoding: 'base58',
+  },
+};
+
 module.exports = DocumentDeleteTransition;
