+++ lib/document/stateTransition/documentTransition/DocumentReplaceTransition.js
@@ -175,4 +175,11 @@ class DocumentReplaceTransition extends AbstractDocumentTransition {
  * @property {number} [$updatedAt]
  */
 
+DocumentReplaceTransition.ENCODED_PROPERTIES = {
+  ...AbstractDocumentTransition.ENCODED_PROPERTIES,
+  $id: {
+    contentEncoding: 'base58',
+  },
+};
+
 module.exports = DocumentReplaceTransition;
