+++ lib/document/stateTransition/validation/data/validateDocumentsBatchTransitionDataFactory.js
@@ -85,7 +85,7 @@ function validateDocumentsBatchTransitionDataFactory(
     documentTransitions
       .forEach((documentTransition) => {
         const fetchedDocument = fetchedDocuments
-          .find((d) => Buffer.compare(documentTransition.getId(), d.getId()) === 0);
+          .find((d) => documentTransition.getId().equals(d.getId()));
 
         switch (documentTransition.getAction()) {
           case AbstractDocumentTransition.ACTIONS.CREATE:
@@ -165,7 +165,7 @@ function validateDocumentsBatchTransitionDataFactory(
               break;
             }
 
-            if (Buffer.compare(fetchedDocument.getOwnerId(), ownerId) !== 0) {
+            if (!fetchedDocument.getOwnerId().equals(ownerId)) {
               result.addError(
                 new DocumentOwnerIdMismatchError(documentTransition, fetchedDocument),
               );
