+++ lib/document/stateTransition/validation/structure/validateDocumentsBatchTransitionStructureFactory.js
@@ -23,6 +23,9 @@ const replaceTransitionSchema = require('../../../../../schema/document/stateTra
 const generateDocumentId = require('../../../generateDocumentId');
 const entropy = require('../../../../util/entropy');
 const EncodedBuffer = require('../../../../util/encoding/EncodedBuffer');
+const encodeObjectProperties = require('../../../../util/encoding/encodeObjectProperties');
+
+const documentsBatchTransitionSchema = require('../../../../../schema/document/stateTransition/documentsBatch');
 
 /**
  * @param {findDuplicatesById} findDuplicatesById
@@ -30,7 +33,7 @@ const EncodedBuffer = require('../../../../util/encoding/EncodedBuffer');
  * @param {validateStateTransitionSignature} validateStateTransitionSignature
  * @param {validateIdentityExistence} validateIdentityExistence
  * @param {StateRepository} stateRepository
- * @param {JsonSchemaValidator} validator
+ * @param {JsonSchemaValidator} jsonSchemaValidator
  * @param {enrichDataContractWithBaseSchema} enrichDataContractWithBaseSchema
  *
  * @return {validateDocumentsBatchTransitionStructure}
@@ -41,7 +44,7 @@ function validateDocumentsBatchTransitionStructureFactory(
   validateStateTransitionSignature,
   validateIdentityExistence,
   stateRepository,
-  validator,
+  jsonSchemaValidator,
   enrichDataContractWithBaseSchema,
 ) {
   const { ACTIONS } = AbstractDocumentTransition;
@@ -122,7 +125,7 @@ function validateDocumentsBatchTransitionStructureFactory(
             enrichedDataContract.toJSON(),
           };
 
-          const schemaResult = validator.validate(
+          const schemaResult = jsonSchemaValidator.validate(
             documentSchemaRef,
             rawDocumentTransition,
             additionalSchemas,
@@ -141,22 +144,17 @@ function validateDocumentsBatchTransitionStructureFactory(
               dataContract.getId(),
               ownerId,
               rawDocumentTransition.$type,
-              bs58.decode(rawDocumentTransition.$entropy),
+              rawDocumentTransition.$entropy,
             );
 
-            const idBuffer = EncodedBuffer.from(
-              rawDocumentTransition.$id,
-              EncodedBuffer.ENCODING.BASE58,
-            );
-
-            if (Buffer.compare(idBuffer, documentId) !== 0) {
+            if (!rawDocumentTransition.$id.equals(documentId)) {
               result.addError(
                 new InvalidDocumentTransitionIdError(rawDocumentTransition),
               );
             }
 
             // validate entropy
-            if (!entropy.validate(bs58.decode(rawDocumentTransition.$entropy))) {
+            if (!entropy.validate(rawDocumentTransition.$entropy)) {
               result.addError(
                 new InvalidDocumentTransitionEntropyError(rawDocumentTransition),
               );
@@ -167,7 +165,7 @@ function validateDocumentsBatchTransitionStructureFactory(
         }
         case ACTIONS.DELETE:
           result.merge(
-            validator.validate(
+            jsonSchemaValidator.validate(
               baseTransitionSchema,
               rawDocumentTransition,
             ),
@@ -213,20 +211,29 @@ function validateDocumentsBatchTransitionStructureFactory(
 
   /**
    * @typedef validateDocumentsBatchTransitionStructure
-   * @param {RawDocumentsBatchTransition} stateTransitionJson
+   * @param {RawDocumentsBatchTransition} rawStateTransition
    * @return {ValidationResult}
    */
-  async function validateDocumentsBatchTransitionStructure(stateTransitionJson) {
-    const result = new ValidationResult();
+  async function validateDocumentsBatchTransitionStructure(rawStateTransition) {
+    // Validate state transition against JSON Schema
+    const jsonStateTransition = encodeObjectProperties(
+      rawStateTransition,
+      DocumentsBatchTransition.ENCODED_PROPERTIES,
+    );
 
-    const ownerIdBuffer = EncodedBuffer.from(
-      stateTransitionJson.ownerId,
-      EncodedBuffer.ENCODING.BASE58,
-    ).toBuffer();
+    const result = jsonSchemaValidator.validate(
+      documentsBatchTransitionSchema,
+      jsonStateTransition,
+    );
+
+    if (!result.isValid()) {
+      return result;
+    }
+
     // Group document transitions by data contracts
-    const documentTransitionsByContracts = stateTransitionJson.transitions
+    const documentTransitionsByContracts = rawStateTransition.transitions
       .reduce((obj, rawDocumentTransition) => {
-        if (!Object.prototype.hasOwnProperty.call(rawDocumentTransition, '$dataContractId')) {
+        if (!Object.prototype.hasOwnProperty.call(rawDocumentTransition, '$dataContractIdString')) {
           result.addError(
             new MissingDataContractIdError(rawDocumentTransition),
           );
@@ -234,7 +241,7 @@ function validateDocumentsBatchTransitionStructureFactory(
           return obj;
         }
 
-        if (typeof rawDocumentTransition.$dataContractId !== 'string') {
+        if (!Buffer.isBuffer(rawDocumentTransition.$dataContractId)) {
           result.addError(
             new InvalidDataContractIdError(rawDocumentTransition.$dataContractId),
           );
@@ -242,17 +249,14 @@ function validateDocumentsBatchTransitionStructureFactory(
           return obj;
         }
 
-        const dataContractId = EncodedBuffer.from(
-          rawDocumentTransition.$dataContractId,
-          EncodedBuffer.ENCODING.BASE58,
-        ).toString();
+        const dataContractIdString = rawDocumentTransition.$dataContractId.toString();
 
-        if (!obj[dataContractId]) {
+        if (!obj[dataContractIdString]) {
           // eslint-disable-next-line no-param-reassign
-          obj[dataContractId] = [];
+          obj[dataContractIdString] = [];
         }
 
-        obj[dataContractId].push(rawDocumentTransition);
+        obj[dataContractIdString].push(rawDocumentTransition);
 
         return obj;
       }, {});
@@ -260,13 +264,10 @@ function validateDocumentsBatchTransitionStructureFactory(
     const dataContracts = [];
 
     const documentTransitionResultsPromises = Object.entries(documentTransitionsByContracts)
-      .map(async ([dataContractId, documentTransitions]) => {
+      .map(async ([dataContractIdString, documentTransitions]) => {
         const perDocumentResult = new ValidationResult();
 
-        const dataContractIdBuffer = EncodedBuffer.from(
-          dataContractId,
-          EncodedBuffer.ENCODING.BASE58,
-        ).toBuffer();
+        const dataContractIdBuffer = Buffer.from(dataContractIdString);
 
         const dataContract = await stateRepository.fetchDataContract(dataContractIdBuffer);
 
@@ -285,7 +286,7 @@ function validateDocumentsBatchTransitionStructureFactory(
         perDocumentResult.merge(
           await validateDocumentTransitions(
             dataContract,
-            ownerIdBuffer,
+            rawStateTransition.ownerId,
             documentTransitions,
           ),
         );
@@ -300,12 +301,10 @@ function validateDocumentsBatchTransitionStructureFactory(
       return result;
     }
 
-    const stateTransition = DocumentsBatchTransition.fromJSON(stateTransitionJson, dataContracts);
-
     // User must exist and confirmed
     result.merge(
       await validateIdentityExistence(
-        stateTransition.getOwnerId(),
+        rawStateTransition.ownerId,
       ),
     );
 
@@ -314,8 +313,13 @@ function validateDocumentsBatchTransitionStructureFactory(
     }
 
     // Verify ST signature
+    const stateTransition = new DocumentsBatchTransition(rawStateTransition);
+
     result.merge(
-      await validateStateTransitionSignature(stateTransition, ownerIdBuffer),
+      await validateStateTransitionSignature(
+        stateTransition,
+        stateTransition.getOwnerId(),
+      ),
     );
 
     return result;
