+++ lib/stateTransition/AbstractStateTransition.js
@@ -28,18 +28,15 @@ class AbstractStateTransition {
    * } [rawStateTransition]
    */
   constructor(rawStateTransition = {}) {
-    this.signature = null;
-    this.protocolVersion = null;
+    this.protocolVersion = rawStateTransition.protocolVersion;
 
-    if (Object.prototype.hasOwnProperty.call(rawStateTransition, 'signature') && rawStateTransition.signature) {
+    this.signature = null;
+
+    if (Object.prototype.hasOwnProperty.call(rawStateTransition, 'signature')) {
       this.signature = EncodedBuffer.from(
         rawStateTransition.signature,
         EncodedBuffer.ENCODING.BASE64,
       );
-    }
-
-    if (Object.prototype.hasOwnProperty.call(rawStateTransition, 'protocolVersion')) {
-      this.protocolVersion = rawStateTransition.protocolVersion;
     }
   }
 
@@ -86,7 +83,7 @@ class AbstractStateTransition {
    * @param {boolean} [options.skipSignature=false]
    * @param {boolean} [options.encodedBuffer=false]
    *
-   * @return {Object}
+   * @return {RawStateTransition}
    */
   toObject(options = {}) {
     Object.assign(
@@ -98,26 +95,20 @@ class AbstractStateTransition {
       },
     );
 
-    let plainObject = {
+    const rawStateTransition = {
       protocolVersion: this.getProtocolVersion(),
       type: this.getType(),
     };
 
     if (!options.skipSignature) {
-      plainObject = {
-        ...plainObject,
-        signature: this.getSignature(),
-      };
+      rawStateTransition.signature = this.getSignature();
     }
 
-    if (!options.encodedBuffer && plainObject.signature) {
-      plainObject = {
-        ...plainObject,
-        signature: this.getSignature().toBuffer(),
-      };
+    if (!options.encodedBuffer && rawStateTransition.signature) {
+      rawStateTransition.signature = this.getSignature().toBuffer();
     }
 
-    return plainObject;
+    return rawStateTransition;
   }
 
   /**
@@ -125,23 +116,20 @@ class AbstractStateTransition {
    *
    * @param {Object} [options]
    *
-   * @return {Object}
+   * @return {JsonStateTransition}
    */
   toJSON(options = {}) {
-    let rawStateTransition = this.toObject({
+    const jsonStateTransition = this.toObject({
       ...options,
       encodedBuffer: true,
     });
 
-    if (!options.skipSignature) {
-      rawStateTransition = {
-        ...rawStateTransition,
-        signature: rawStateTransition.signature === null
-          ? null : rawStateTransition.signature.toString(),
-      };
+    if (!options.skipSignature && jsonStateTransition.signature !== null) {
+      // noinspection JSValidateTypes
+      jsonStateTransition.signature = jsonStateTransition.signature.toString();
     }
 
-    return rawStateTransition;
+    return jsonStateTransition;
   }
 
   /**
@@ -229,7 +217,20 @@ class AbstractStateTransition {
  * @typedef RawStateTransition
  * @property {number} protocolVersion
  * @property {number} type
+ * @property {Buffer|null} signature
+ */
+
+/**
+ * @typedef JsonStateTransition
+ * @property {number} protocolVersion
+ * @property {number} type
  * @property {string|null} signature
  */
+
+AbstractStateTransition.ENCODED_PROPERTIES = {
+  signature: {
+    contentEncoding: 'base64',
+  },
+};
 
 module.exports = AbstractStateTransition;
