+++ lib/stateTransition/AbstractStateTransitionIdentitySigned.js
@@ -127,16 +127,13 @@ class AbstractStateTransitionIdentitySigned extends AbstractStateTransition {
   toObject(options = {}) {
     const skipSignature = !!options.skipSignature;
 
-    let plainObject = super.toObject(options);
+    const rawStateTransition = super.toObject(options);
 
     if (!skipSignature) {
-      plainObject = {
-        ...plainObject,
-        signaturePublicKeyId: this.getSignaturePublicKeyId(),
-      };
+      rawStateTransition.signaturePublicKeyId = this.getSignaturePublicKeyId();
     }
 
-    return plainObject;
+    return rawStateTransition;
   }
 
   /**
@@ -150,5 +147,9 @@ class AbstractStateTransitionIdentitySigned extends AbstractStateTransition {
     return AbstractStateTransition.translateJsonToObject(rawStateTransition);
   }
 }
+
+AbstractStateTransitionIdentitySigned.ENCODED_PROPERTIES = {
+  ...AbstractStateTransition.ENCODED_PROPERTIES,
+};
 
 module.exports = AbstractStateTransitionIdentitySigned;
