+++ lib/stateTransition/StateTransitionFacade.js
@@ -9,8 +9,6 @@ const AbstractStateTransition = require('./AbstractStateTransition');
 
 const stateTransitionTypes = require('./stateTransitionTypes');
 
-const dataContractCreateTransitionSchema = require('../../schema/dataContract/stateTransition/dataContractCreate');
-const documentsBatchTransitionSchema = require('../../schema/document/stateTransition/documentsBatch');
 const identityCreateTransitionSchema = require('../../schema/identity/stateTransition/identityCreate');
 const identityTopUpTransitionSchema = require('../../schema/identity/stateTransition/identityTopUp');
 
@@ -117,28 +115,16 @@ class StateTransitionFacade {
       validatePublicKeys,
     );
 
-    const typeExtensions = {
-      [stateTransitionTypes.DATA_CONTRACT_CREATE]: {
-        validationFunction: validateDataContractCreateTransitionStructure,
-        schema: dataContractCreateTransitionSchema,
-      },
-      [stateTransitionTypes.DOCUMENTS_BATCH]: {
-        validationFunction: validateDocumentsBatchTransitionStructure,
-        schema: documentsBatchTransitionSchema,
-      },
-      [stateTransitionTypes.IDENTITY_CREATE]: {
-        validationFunction: validateIdentityCreateSTStructure,
-        schema: identityCreateTransitionSchema,
-      },
-      [stateTransitionTypes.IDENTITY_TOP_UP]: {
-        validationFunction: validateIdentityTopUpTransitionStructure,
-        schema: identityTopUpTransitionSchema,
-      },
+    const validationFunctionsByType = {
+      [stateTransitionTypes.DATA_CONTRACT_CREATE]: validateDataContractCreateTransitionStructure,
+      [stateTransitionTypes.DOCUMENTS_BATCH]: validateDocumentsBatchTransitionStructure,
+      [stateTransitionTypes.IDENTITY_CREATE]: validateIdentityCreateSTStructure,
+      [stateTransitionTypes.IDENTITY_TOP_UP]: validateIdentityTopUpTransitionStructure,
     };
 
     this.validateStateTransitionStructure = validateStateTransitionStructureFactory(
       validator,
-      typeExtensions,
+      validationFunctionsByType,
       this.createStateTransition,
     );
 
@@ -311,12 +297,7 @@ class StateTransitionFacade {
   /**
    * Validate State Transition Structure
    *
-   * @param {
-   *  DataContractCreateTransition
-   * |RawDataContractCreateTransition
-   * |RawDocumentsBatchTransition
-   * |DocumentsBatchTransition
-   * } stateTransition
+   * @param {AbstractStateTransition} stateTransition
    * @return {ValidationResult}
    */
   async validateStructure(stateTransition) {
@@ -328,17 +309,13 @@ class StateTransitionFacade {
       );
     }
 
-    return this.validateStateTransitionStructure(stateTransition);
+    return this.validateStateTransitionStructure(stateTransition.toObject());
   }
 
   /**
    * Validate State Transition Data
    *
-   * @param {
-   *  DataContractCreateTransition|DocumentsBatchTransition|IdentityCreateTransition
-   *  |RawDataContractCreateTransition|RawDocumentsBatchTransition|RawIdentityCreateTransition
-   *  |RawIdentityTopUpTransition
-   *  } stateTransition
+   * @param {AbstractStateTransition} stateTransition
    * @return {ValidationResult}
    */
   async validateData(stateTransition) {
@@ -349,17 +326,14 @@ class StateTransitionFacade {
         + ' setStateRepository method',
       );
     }
-    let stateTransitionModel = stateTransition;
 
-    if (!(stateTransition instanceof AbstractStateTransition)) {
-      stateTransitionModel = await this.createFromObject(stateTransition);
-    }
-
-    return this.validateStateTransitionData(stateTransitionModel);
+    return this.validateStateTransitionData(stateTransition);
   }
 
   /**
    * Validate State Transition Fee
+   *
+   * @param {AbstractStateTransition} stateTransition
    *
    * @return {ValidationResult}
    */
@@ -372,13 +346,7 @@ class StateTransitionFacade {
       );
     }
 
-    let stateTransitionModel = stateTransition;
-
-    if (!(stateTransition instanceof AbstractStateTransition)) {
-      stateTransitionModel = await this.createFromObject(stateTransition);
-    }
-
-    return this.validateStateTransitionFee(stateTransitionModel);
+    return this.validateStateTransitionFee(stateTransition);
   }
 
   /**
