+++ lib/stateTransition/StateTransitionFactory.js
@@ -47,36 +47,26 @@ class StateTransitionFactory {
   /**
    * Create State Transition from plain object
    *
-   * @param {RawDataContractCreateTransition|RawDocumentsBatchTransition} rawStateTransition
-   * @param {Object} options
+   * @param {RawStateTransition} rawStateTransition
+   * @param {Object} [options]
    * @param {boolean} [options.skipValidation=false]
-   * @return {RawDataContractCreateTransition|DocumentsBatchTransition}
+   * @return {AbstractStateTransition}
    */
   async createFromObject(rawStateTransition, options = {}) {
     const opts = { skipValidation: false, ...options };
 
-    // noinspection UnnecessaryLocalVariableJS
-    const stateTransition = await this.createStateTransition(rawStateTransition, {
-      fromJSON: false,
-    });
-
     if (!opts.skipValidation) {
-      let stateTransitionJson;
-      try {
-        stateTransitionJson = stateTransition.toJSON();
-      } catch (e) {
-        if (e instanceof ConsensusError) {
-          throw new InvalidStateTransitionError([e], rawStateTransition);
-        }
-        throw e;
-      }
-
-      const result = await this.validateStateTransitionStructure(stateTransitionJson);
+      const result = await this.validateStateTransitionStructure(rawStateTransition);
 
       if (!result.isValid()) {
         throw new InvalidStateTransitionError(result.getErrors(), rawStateTransition);
       }
     }
+
+    // noinspection UnnecessaryLocalVariableJS
+    const stateTransition = await this.createStateTransition(rawStateTransition, {
+      fromJSON: false,
+    });
 
     return stateTransition;
   }
