+++ lib/util/encoding/EncodedBuffer.js
@@ -64,7 +64,11 @@ EncodedBuffer.prototype.toBuffer = function toBuffer() {
  *
  * @return {string}
  */
-EncodedBuffer.prototype.toString = function toString() {
+EncodedBuffer.prototype.toString = function toString(variableEncoding) {
+  if (variableEncoding) {
+    throw new TypeError('Encoding is preset already');
+  }
+
   const encoding = this.getEncoding();
 
   switch (this.getEncoding()) {
