+++ schema/dataContract/stateTransition/dataContractCreate.json
@@ -7,6 +7,10 @@
       "maximum": 0,
       "$comment": "Maximum is the latest Data Contract Create Transition protocol version"
     },
+    "type": {
+      "type": "integer",
+      "const": 0
+    },
     "dataContract": {
       "type": "object"
     },
@@ -16,11 +20,26 @@
       "minLength": 26,
       "maxLength": 35,
       "pattern": "^[123456789ABCDEFGHJKLMNPQRSTUVWXYZabcdefghijkmnopqrstuvwxyz]+$"
+    },
+    "signaturePublicKeyId": {
+      "type": "integer",
+      "minimum": 0
+    },
+    "signature": {
+      "type": "string",
+      "contentEncoding": "base64",
+      "minLength": 86,
+      "maxLength": 86,
+      "pattern": "^([A-Za-z0-9+/])+$"
     }
   },
+  "additionalProperties": false,
   "required": [
     "protocolVersion",
+    "type",
     "dataContract",
-    "entropy"
+    "entropy",
+    "signaturePublicKeyId",
+    "signature"
   ]
 }
