--- schema/stateTransition/stateTransitionBase.json
@@ -1,25 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-07/schema",
-  "properties": {
-    "type": {
-      "type": "integer",
-      "enum": [0, 1, 2, 3]
-    },
-    "signaturePublicKeyId": {
-      "type": ["integer", "null"],
-      "minimum": 0
-    },
-    "signature": {
-      "type": "string",
-      "contentEncoding": "base64",
-      "minLength": 86,
-      "maxLength": 88,
-      "pattern": "^([A-Za-z0-9+/])+$"
-    }
-  },
-  "required": [
-    "type",
-    "signature"
-  ],
-  "additionalProperties": false
-}
