+++ test/integration/dataContract/validateDataContractFactory.spec.js
@@ -26,11 +26,9 @@ describe('validateDataContractFactory', () => {
   let rawDataContract;
   let validateDataContract;
 
-  beforeEach(function beforeEach() {
+  beforeEach(() => {
     dataContract = getDataContractFixture();
-    rawDataContract = dataContract.toJSON();
-    dataContract.toJSON = this.sinonSandbox.stub()
-      .returns(rawDataContract);
+    rawDataContract = dataContract.toObject();
 
     const ajv = new Ajv();
     const jsonSchemaValidator = new JsonSchemaValidator(ajv);
@@ -48,7 +46,7 @@ describe('validateDataContractFactory', () => {
     it('should be present', async () => {
       delete rawDataContract.protocolVersion;
 
-      const result = await validateDataContract(dataContract);
+      const result = await validateDataContract(rawDataContract);
 
       expectJsonSchemaError(result);
 
@@ -62,7 +60,7 @@ describe('validateDataContractFactory', () => {
     it('should be an integer', async () => {
       rawDataContract.protocolVersion = '1';
 
-      const result = await validateDataContract(dataContract);
+      const result = await validateDataContract(rawDataContract);
 
       expectJsonSchemaError(result);
 
@@ -75,7 +73,7 @@ describe('validateDataContractFactory', () => {
     it('should not be less than 0', async () => {
       rawDataContract.protocolVersion = -1;
 
-      const result = await validateDataContract(dataContract);
+      const result = await validateDataContract(rawDataContract);
 
       expectJsonSchemaError(result);
 
@@ -88,7 +86,7 @@ describe('validateDataContractFactory', () => {
     it('should not be greater than current version (0)', async () => {
       rawDataContract.protocolVersion = 1;
 
-      const result = await validateDataContract(dataContract);
+      const result = await validateDataContract(rawDataContract);
 
       expectJsonSchemaError(result);
 
@@ -103,7 +101,7 @@ describe('validateDataContractFactory', () => {
     it('should be present', async () => {
       delete rawDataContract.$schema;
 
-      const result = await validateDataContract(dataContract);
+      const result = await validateDataContract(rawDataContract);
 
       expectJsonSchemaError(result);
 
@@ -117,7 +115,7 @@ describe('validateDataContractFactory', () => {
     it('should be a string', async () => {
       rawDataContract.$schema = 1;
 
-      const result = await validateDataContract(dataContract);
+      const result = await validateDataContract(rawDataContract);
 
       expectJsonSchemaError(result);
 
@@ -130,7 +128,7 @@ describe('validateDataContractFactory', () => {
     it('should be a particular url', async () => {
       rawDataContract.$schema = 'wrong';
 
-      const result = await validateDataContract(dataContract);
+      const result = await validateDataContract(rawDataContract);
 
       expectJsonSchemaError(result);
 
@@ -145,7 +143,7 @@ describe('validateDataContractFactory', () => {
     it('should be present', async () => {
       delete rawDataContract.ownerId;
 
-      const result = await validateDataContract(dataContract);
+      const result = await validateDataContract(rawDataContract);
 
       expectJsonSchemaError(result);
 
@@ -159,7 +157,7 @@ describe('validateDataContractFactory', () => {
     it('should be a string', async () => {
       rawDataContract.ownerId = 1;
 
-      const result = await validateDataContract(dataContract);
+      const result = await validateDataContract(rawDataContract);
 
       expectJsonSchemaError(result);
 
@@ -172,7 +170,7 @@ describe('validateDataContractFactory', () => {
     it('should be no less than 42 chars', async () => {
       rawDataContract.ownerId = '1'.repeat(41);
 
-      const result = await validateDataContract(dataContract);
+      const result = await validateDataContract(rawDataContract);
 
       expectJsonSchemaError(result);
 
@@ -185,7 +183,7 @@ describe('validateDataContractFactory', () => {
     it('should be no longer than 44 chars', async () => {
       rawDataContract.ownerId = '1'.repeat(45);
 
-      const result = await validateDataContract(dataContract);
+      const result = await validateDataContract(rawDataContract);
 
       expectJsonSchemaError(result);
 
@@ -198,7 +196,7 @@ describe('validateDataContractFactory', () => {
     it('should be base58 encoded', async () => {
       rawDataContract.ownerId = '&'.repeat(44);
 
-      const result = await validateDataContract(dataContract);
+      const result = await validateDataContract(rawDataContract);
 
       expectJsonSchemaError(result);
 
@@ -213,7 +211,7 @@ describe('validateDataContractFactory', () => {
     it('should be present', async () => {
       delete rawDataContract.$id;
 
-      const result = await validateDataContract(dataContract);
+      const result = await validateDataContract(rawDataContract);
 
       expectJsonSchemaError(result);
 
@@ -227,7 +225,7 @@ describe('validateDataContractFactory', () => {
     it('should be a string', async () => {
       rawDataContract.$id = 1;
 
-      const result = await validateDataContract(dataContract);
+      const result = await validateDataContract(rawDataContract);
 
       expectJsonSchemaError(result);
 
@@ -240,7 +238,7 @@ describe('validateDataContractFactory', () => {
     it('should be no less than 42 chars', async () => {
       rawDataContract.$id = '1'.repeat(41);
 
-      const result = await validateDataContract(dataContract);
+      const result = await validateDataContract(rawDataContract);
 
       expectJsonSchemaError(result);
 
@@ -253,7 +251,7 @@ describe('validateDataContractFactory', () => {
     it('should be no longer than 44 chars', async () => {
       rawDataContract.$id = '1'.repeat(45);
 
-      const result = await validateDataContract(dataContract);
+      const result = await validateDataContract(rawDataContract);
 
       expectJsonSchemaError(result);
 
@@ -266,7 +264,7 @@ describe('validateDataContractFactory', () => {
     it('should be base58 encoded', async () => {
       rawDataContract.$id = '&'.repeat(44);
 
-      const result = await validateDataContract(dataContract);
+      const result = await validateDataContract(rawDataContract);
 
       expectJsonSchemaError(result);
 
@@ -282,7 +280,7 @@ describe('validateDataContractFactory', () => {
       delete rawDataContract.definitions;
       delete rawDataContract.documents.prettyDocument;
 
-      const result = await validateDataContract(dataContract);
+      const result = await validateDataContract(rawDataContract);
 
       expect(result).to.be.an.instanceOf(ValidationResult);
       expect(result.isValid()).to.be.true();
@@ -291,7 +289,7 @@ describe('validateDataContractFactory', () => {
     it('should be an object', async () => {
       rawDataContract.definitions = 1;
 
-      const result = await validateDataContract(dataContract);
+      const result = await validateDataContract(rawDataContract);
 
       expectJsonSchemaError(result);
 
@@ -304,7 +302,7 @@ describe('validateDataContractFactory', () => {
     it('should not be empty', async () => {
       rawDataContract.definitions = {};
 
-      const result = await validateDataContract(dataContract);
+      const result = await validateDataContract(rawDataContract);
 
       expectJsonSchemaError(result);
 
@@ -319,7 +317,7 @@ describe('validateDataContractFactory', () => {
         $subSchema: {},
       };
 
-      const result = await validateDataContract(dataContract);
+      const result = await validateDataContract(rawDataContract);
 
       expectJsonSchemaError(result, 2);
 
@@ -339,7 +337,7 @@ describe('validateDataContractFactory', () => {
         rawDataContract.definitions[i] = item;
       });
 
-      const result = await validateDataContract(dataContract);
+      const result = await validateDataContract(rawDataContract);
 
       expectJsonSchemaError(result);
 
@@ -359,7 +357,7 @@ describe('validateDataContractFactory', () => {
             type: 'string',
           };
 
-          const result = await validateDataContract(dataContract);
+          const result = await validateDataContract(rawDataContract);
 
           expectJsonSchemaError(result, 0);
         }),
@@ -375,7 +373,7 @@ describe('validateDataContractFactory', () => {
             type: 'string',
           };
 
-          const result = await validateDataContract(dataContract);
+          const result = await validateDataContract(rawDataContract);
 
           expectJsonSchemaError(result, 2);
 
@@ -392,7 +390,7 @@ describe('validateDataContractFactory', () => {
     it('should be present', async () => {
       delete rawDataContract.documents;
 
-      const result = await validateDataContract(dataContract);
+      const result = await validateDataContract(rawDataContract);
 
       expectJsonSchemaError(result);
 
@@ -406,7 +404,7 @@ describe('validateDataContractFactory', () => {
     it('should be an object', async () => {
       rawDataContract.documents = 1;
 
-      const result = await validateDataContract(dataContract);
+      const result = await validateDataContract(rawDataContract);
 
       expectJsonSchemaError(result);
 
@@ -419,7 +417,7 @@ describe('validateDataContractFactory', () => {
     it('should not be empty', async () => {
       rawDataContract.documents = {};
 
-      const result = await validateDataContract(dataContract);
+      const result = await validateDataContract(rawDataContract);
 
       expectJsonSchemaError(result);
 
@@ -437,7 +435,7 @@ describe('validateDataContractFactory', () => {
         validNames.map(async (name) => {
           rawDataContract.documents[name] = rawDataContract.documents.niceDocument;
 
-          const result = await validateDataContract(dataContract);
+          const result = await validateDataContract(rawDataContract);
 
           expectJsonSchemaError(result, 0);
         }),
@@ -451,7 +449,7 @@ describe('validateDataContractFactory', () => {
         invalidNames.map(async (name) => {
           rawDataContract.documents[name] = rawDataContract.documents.niceDocument;
 
-          const result = await validateDataContract(dataContract);
+          const result = await validateDataContract(rawDataContract);
 
           expectJsonSchemaError(result, 2);
 
@@ -472,7 +470,7 @@ describe('validateDataContractFactory', () => {
         rawDataContract.documents[i] = item;
       });
 
-      const result = await validateDataContract(dataContract);
+      const result = await validateDataContract(rawDataContract);
 
       expectJsonSchemaError(result);
 
@@ -486,7 +484,7 @@ describe('validateDataContractFactory', () => {
       it('should not be empty', async () => {
         rawDataContract.documents.niceDocument.properties = {};
 
-        const result = await validateDataContract(dataContract);
+        const result = await validateDataContract(rawDataContract);
 
         expectJsonSchemaError(result);
 
@@ -499,7 +497,7 @@ describe('validateDataContractFactory', () => {
       it('should have type "object"', async () => {
         rawDataContract.documents.niceDocument.type = 'string';
 
-        const result = await validateDataContract(dataContract);
+        const result = await validateDataContract(rawDataContract);
 
         expectJsonSchemaError(result);
 
@@ -512,7 +510,7 @@ describe('validateDataContractFactory', () => {
       it('should have "properties"', async () => {
         delete rawDataContract.documents.niceDocument.properties;
 
-        const result = await validateDataContract(dataContract);
+        const result = await validateDataContract(rawDataContract);
 
         expectJsonSchemaError(result);
 
@@ -538,7 +536,7 @@ describe('validateDataContractFactory', () => {
           ],
         };
 
-        const result = await validateDataContract(dataContract);
+        const result = await validateDataContract(rawDataContract);
 
         expectJsonSchemaError(result, 3);
 
@@ -559,7 +557,7 @@ describe('validateDataContractFactory', () => {
               type: 'string',
             };
 
-            const result = await validateDataContract(dataContract);
+            const result = await validateDataContract(rawDataContract);
 
             expectJsonSchemaError(result, 0);
           }),
@@ -581,7 +579,7 @@ describe('validateDataContractFactory', () => {
               type: 'string',
             };
 
-            const result = await validateDataContract(dataContract);
+            const result = await validateDataContract(rawDataContract);
 
             expectJsonSchemaError(result, 0);
           }),
@@ -595,7 +593,7 @@ describe('validateDataContractFactory', () => {
           invalidNames.map(async (name) => {
             rawDataContract.documents.niceDocument.properties[name] = {};
 
-            const result = await validateDataContract(dataContract);
+            const result = await validateDataContract(rawDataContract);
 
             expectJsonSchemaError(result, 2);
 
@@ -621,7 +619,7 @@ describe('validateDataContractFactory', () => {
           invalidNames.map(async (name) => {
             rawDataContract.documents.niceDocument.properties.something.properties[name] = {};
 
-            const result = await validateDataContract(dataContract);
+            const result = await validateDataContract(rawDataContract);
 
             expectJsonSchemaError(result, 2);
 
@@ -642,7 +640,7 @@ describe('validateDataContractFactory', () => {
       it('should have "additionalProperties" defined', async () => {
         delete rawDataContract.documents.niceDocument.additionalProperties;
 
-        const result = await validateDataContract(dataContract);
+        const result = await validateDataContract(rawDataContract);
 
         expectJsonSchemaError(result);
 
@@ -656,7 +654,7 @@ describe('validateDataContractFactory', () => {
       it('should have "additionalProperties" defined to false', async () => {
         rawDataContract.documents.niceDocument.additionalProperties = true;
 
-        const result = await validateDataContract(dataContract);
+        const result = await validateDataContract(rawDataContract);
 
         expectJsonSchemaError(result);
 
@@ -680,7 +678,7 @@ describe('validateDataContractFactory', () => {
           ],
         };
 
-        const result = await validateDataContract(dataContract);
+        const result = await validateDataContract(rawDataContract);
 
         expectJsonSchemaError(result, 3);
 
@@ -694,7 +692,7 @@ describe('validateDataContractFactory', () => {
       it('should return invalid result if there are additional properties', async () => {
         rawDataContract.additionalProperty = { };
 
-        const result = await validateDataContract(dataContract);
+        const result = await validateDataContract(rawDataContract);
 
         expectJsonSchemaError(result);
 
@@ -713,7 +711,7 @@ describe('validateDataContractFactory', () => {
           rawDataContract.documents.niceDocument.properties[i] = item;
         });
 
-        const result = await validateDataContract(dataContract);
+        const result = await validateDataContract(rawDataContract);
 
         expectJsonSchemaError(result);
 
@@ -733,7 +731,7 @@ describe('validateDataContractFactory', () => {
           additionalProperties: false,
         };
 
-        const result = await validateDataContract(dataContract);
+        const result = await validateDataContract(rawDataContract);
 
         expectJsonSchemaError(result);
 
@@ -757,7 +755,7 @@ describe('validateDataContractFactory', () => {
           additionalProperties: false,
         };
 
-        const result = await validateDataContract(dataContract);
+        const result = await validateDataContract(rawDataContract);
 
         expectJsonSchemaError(result, 0);
       });
@@ -780,7 +778,7 @@ describe('validateDataContractFactory', () => {
           additionalProperties: false,
         };
 
-        const result = await validateDataContract(dataContract);
+        const result = await validateDataContract(rawDataContract);
 
         expectJsonSchemaError(result);
 
@@ -810,7 +808,7 @@ describe('validateDataContractFactory', () => {
           additionalProperties: false,
         };
 
-        const result = await validateDataContract(dataContract);
+        const result = await validateDataContract(rawDataContract);
 
         expectJsonSchemaError(result, 0);
       });
@@ -834,7 +832,7 @@ describe('validateDataContractFactory', () => {
           additionalProperties: false,
         };
 
-        const result = await validateDataContract(dataContract);
+        const result = await validateDataContract(rawDataContract);
 
         expectJsonSchemaError(result, 3);
 
@@ -854,7 +852,7 @@ describe('validateDataContractFactory', () => {
       it('should return invalid result if "default" keyword is used', async () => {
         rawDataContract.documents.indexedDocument.properties.firstName.default = '1';
 
-        const result = await validateDataContract(dataContract);
+        const result = await validateDataContract(rawDataContract);
 
         expectJsonSchemaError(result);
 
@@ -869,7 +867,7 @@ describe('validateDataContractFactory', () => {
           $ref: 'http://remote.com/schema#',
         };
 
-        const result = await validateDataContract(dataContract);
+        const result = await validateDataContract(rawDataContract);
 
         expectJsonSchemaError(result);
 
@@ -893,7 +891,7 @@ describe('validateDataContractFactory', () => {
           additionalProperties: false,
         };
 
-        const result = await validateDataContract(dataContract);
+        const result = await validateDataContract(rawDataContract);
 
         expectJsonSchemaError(result);
 
@@ -916,7 +914,7 @@ describe('validateDataContractFactory', () => {
           additionalProperties: false,
         };
 
-        const result = await validateDataContract(dataContract);
+        const result = await validateDataContract(rawDataContract);
 
         expectJsonSchemaError(result);
 
@@ -940,7 +938,7 @@ describe('validateDataContractFactory', () => {
           additionalProperties: false,
         };
 
-        const result = await validateDataContract(dataContract);
+        const result = await validateDataContract(rawDataContract);
 
         expectJsonSchemaError(result);
 
@@ -963,7 +961,7 @@ describe('validateDataContractFactory', () => {
           additionalProperties: false,
         };
 
-        const result = await validateDataContract(dataContract);
+        const result = await validateDataContract(rawDataContract);
 
         expectJsonSchemaError(result);
 
@@ -986,7 +984,7 @@ describe('validateDataContractFactory', () => {
           additionalProperties: false,
         };
 
-        const result = await validateDataContract(dataContract);
+        const result = await validateDataContract(rawDataContract);
 
         expectJsonSchemaError(result);
 
@@ -1010,7 +1008,7 @@ describe('validateDataContractFactory', () => {
           additionalProperties: false,
         };
 
-        const result = await validateDataContract(dataContract);
+        const result = await validateDataContract(rawDataContract);
 
         expectJsonSchemaError(result);
 
@@ -1032,7 +1030,7 @@ describe('validateDataContractFactory', () => {
           additionalProperties: false,
         };
 
-        const result = await validateDataContract(dataContract);
+        const result = await validateDataContract(rawDataContract);
 
         expectJsonSchemaError(result);
 
@@ -1056,7 +1054,7 @@ describe('validateDataContractFactory', () => {
           additionalProperties: false,
         };
 
-        const result = await validateDataContract(dataContract);
+        const result = await validateDataContract(rawDataContract);
 
         expectJsonSchemaError(result);
 
@@ -1069,7 +1067,7 @@ describe('validateDataContractFactory', () => {
       it('should have `pattern` set if `contentEncoding` is set to `base64`', async () => {
         delete rawDataContract.documents.withContentEncoding.properties.base64Field.pattern;
 
-        const result = await validateDataContract(dataContract);
+        const result = await validateDataContract(rawDataContract);
 
         expectJsonSchemaError(result);
 
@@ -1082,7 +1080,7 @@ describe('validateDataContractFactory', () => {
       it('should have specific `pattern` set if `contentEncoding` is set to `base64`', async () => {
         rawDataContract.documents.withContentEncoding.properties.base64Field.pattern = '.*';
 
-        const result = await validateDataContract(dataContract);
+        const result = await validateDataContract(rawDataContract);
 
         expectJsonSchemaError(result);
 
@@ -1097,7 +1095,7 @@ describe('validateDataContractFactory', () => {
       it('should have `pattern` set if `contentEncoding` is set to `base58`', async () => {
         delete rawDataContract.documents.withContentEncoding.properties.base58Field.pattern;
 
-        const result = await validateDataContract(dataContract);
+        const result = await validateDataContract(rawDataContract);
 
         expectJsonSchemaError(result);
 
@@ -1110,7 +1108,7 @@ describe('validateDataContractFactory', () => {
       it('should have specific `pattern` set if `contentEncoding` is set to `base58`', async () => {
         rawDataContract.documents.withContentEncoding.properties.base58Field.pattern = '.*';
 
-        const result = await validateDataContract(dataContract);
+        const result = await validateDataContract(rawDataContract);
 
         expectJsonSchemaError(result);
 
@@ -1125,7 +1123,7 @@ describe('validateDataContractFactory', () => {
       it('should have contentEncoding set to either base64 or base58', async () => {
         rawDataContract.documents.withContentEncoding.properties.base58Field.contentEncoding = 'binary';
 
-        const result = await validateDataContract(dataContract);
+        const result = await validateDataContract(rawDataContract);
 
         expectJsonSchemaError(result);
 
@@ -1143,7 +1141,7 @@ describe('validateDataContractFactory', () => {
     it('should be an array', async () => {
       rawDataContract.documents.indexedDocument.indices = 'definitely not an array';
 
-      const result = await validateDataContract(dataContract);
+      const result = await validateDataContract(rawDataContract);
 
       expectJsonSchemaError(result);
 
@@ -1156,7 +1154,7 @@ describe('validateDataContractFactory', () => {
     it('should have at least one item', async () => {
       rawDataContract.documents.indexedDocument.indices = [];
 
-      const result = await validateDataContract(dataContract);
+      const result = await validateDataContract(rawDataContract);
 
       expectJsonSchemaError(result);
 
@@ -1171,7 +1169,7 @@ describe('validateDataContractFactory', () => {
 
       rawDataContract.documents.indexedDocument.indices.push(indexDefinition);
 
-      const result = await validateDataContract(dataContract);
+      const result = await validateDataContract(rawDataContract);
 
       expectValidationError(result, DuplicateIndexError);
 
@@ -1186,7 +1184,7 @@ describe('validateDataContractFactory', () => {
       it('should be an object', async () => {
         rawDataContract.documents.indexedDocument.indices = ['something else'];
 
-        const result = await validateDataContract(dataContract);
+        const result = await validateDataContract(rawDataContract);
 
         expectJsonSchemaError(result);
 
@@ -1199,7 +1197,7 @@ describe('validateDataContractFactory', () => {
       it('should have properties definition', async () => {
         rawDataContract.documents.indexedDocument.indices = [{}];
 
-        const result = await validateDataContract(dataContract);
+        const result = await validateDataContract(rawDataContract);
 
         expectJsonSchemaError(result);
 
@@ -1215,7 +1213,7 @@ describe('validateDataContractFactory', () => {
           rawDataContract.documents.indexedDocument.indices[0]
             .properties = 'something else';
 
-          const result = await validateDataContract(dataContract);
+          const result = await validateDataContract(rawDataContract);
 
           expectJsonSchemaError(result);
 
@@ -1231,7 +1229,7 @@ describe('validateDataContractFactory', () => {
           rawDataContract.documents.indexedDocument.indices[0]
             .properties = [];
 
-          const result = await validateDataContract(dataContract);
+          const result = await validateDataContract(rawDataContract);
 
           expectJsonSchemaError(result);
 
@@ -1249,7 +1247,7 @@ describe('validateDataContractFactory', () => {
               .properties.push({ [`field${i}`]: 'asc' });
           }
 
-          const result = await validateDataContract(dataContract);
+          const result = await validateDataContract(rawDataContract);
 
           expectJsonSchemaError(result);
 
@@ -1266,7 +1264,7 @@ describe('validateDataContractFactory', () => {
             rawDataContract.documents.indexedDocument.indices[0]
               .properties[0] = 'something else';
 
-            const result = await validateDataContract(dataContract);
+            const result = await validateDataContract(rawDataContract);
 
             expectJsonSchemaError(result);
 
@@ -1282,7 +1280,7 @@ describe('validateDataContractFactory', () => {
             rawDataContract.documents.indexedDocument.indices[0]
               .properties = [];
 
-            const result = await validateDataContract(dataContract);
+            const result = await validateDataContract(rawDataContract);
 
             expectJsonSchemaError(result);
 
@@ -1300,7 +1298,7 @@ describe('validateDataContractFactory', () => {
 
             property.anotherField = 'something';
 
-            const result = await validateDataContract(dataContract);
+            const result = await validateDataContract(rawDataContract);
 
             expectJsonSchemaError(result);
 
@@ -1316,7 +1314,7 @@ describe('validateDataContractFactory', () => {
             rawDataContract.documents.indexedDocument.indices[0]
               .properties[0].$ownerId = 'wrong';
 
-            const result = await validateDataContract(dataContract);
+            const result = await validateDataContract(rawDataContract);
 
             expectJsonSchemaError(result);
 
@@ -1333,7 +1331,7 @@ describe('validateDataContractFactory', () => {
       it('should have "unique" flag to be of a boolean type', async () => {
         rawDataContract.documents.indexedDocument.indices[0].unique = 12;
 
-        const result = await validateDataContract(dataContract);
+        const result = await validateDataContract(rawDataContract);
 
         expectJsonSchemaError(result);
 
@@ -1354,7 +1352,7 @@ describe('validateDataContractFactory', () => {
           });
         }
 
-        const result = await validateDataContract(dataContract);
+        const result = await validateDataContract(rawDataContract);
 
         expectJsonSchemaError(result);
 
@@ -1381,7 +1379,7 @@ describe('validateDataContractFactory', () => {
           });
         }
 
-        const result = await validateDataContract(dataContract);
+        const result = await validateDataContract(rawDataContract);
 
         expectValidationError(result, UniqueIndicesLimitReachedError);
 
@@ -1402,7 +1400,7 @@ describe('validateDataContractFactory', () => {
 
         indeciesDefinition.push(indexDefinition);
 
-        const result = await validateDataContract(dataContract);
+        const result = await validateDataContract(rawDataContract);
 
         expectValidationError(result, SystemPropertyIndexAlreadyPresentError);
 
@@ -1421,7 +1419,7 @@ describe('validateDataContractFactory', () => {
           missingProperty: 'asc',
         });
 
-        const result = await validateDataContract(dataContract);
+        const result = await validateDataContract(rawDataContract);
 
         expectValidationError(result, UndefinedIndexPropertyError);
 
@@ -1454,7 +1452,7 @@ describe('validateDataContractFactory', () => {
           objectProperty: 'asc',
         });
 
-        const result = await validateDataContract(dataContract);
+        const result = await validateDataContract(rawDataContract);
 
         expectValidationError(result, InvalidIndexPropertyTypeError);
 
@@ -1491,7 +1489,7 @@ describe('validateDataContractFactory', () => {
           arrayProperty: 'asc',
         });
 
-        const result = await validateDataContract(dataContract);
+        const result = await validateDataContract(rawDataContract);
 
         expectValidationError(result, InvalidIndexPropertyTypeError);
 
@@ -1525,7 +1523,7 @@ describe('validateDataContractFactory', () => {
           arrayProperty: 'asc',
         });
 
-        const result = await validateDataContract(dataContract);
+        const result = await validateDataContract(rawDataContract);
 
         expectValidationError(result, InvalidIndexPropertyTypeError);
 
@@ -1559,7 +1557,7 @@ describe('validateDataContractFactory', () => {
           arrayProperty: 'asc',
         });
 
-        const result = await validateDataContract(dataContract);
+        const result = await validateDataContract(rawDataContract);
 
         expectValidationError(result, InvalidIndexPropertyTypeError);
 
@@ -1575,7 +1573,7 @@ describe('validateDataContractFactory', () => {
       it('should return invalid result if unique compound index contains both required and optional properties', async () => {
         rawDataContract.documents.optionalUniqueIndexedDocument.required.splice(-1);
 
-        const result = await validateDataContract(dataContract);
+        const result = await validateDataContract(rawDataContract);
 
         expectValidationError(result, InvalidCompoundIndexError);
 
@@ -1592,7 +1590,7 @@ describe('validateDataContractFactory', () => {
   it('should return invalid result with circular $ref pointer', async () => {
     rawDataContract.definitions.object = { $ref: '#/definitions/object' };
 
-    const result = await validateDataContract(dataContract);
+    const result = await validateDataContract(rawDataContract);
 
     expectJsonSchemaError(result);
 
@@ -1606,7 +1604,7 @@ describe('validateDataContractFactory', () => {
   it('should return invalid result if indexed property missing maxLength constraint', async () => {
     delete rawDataContract.documents.indexedDocument.properties.firstName.maxLength;
 
-    const result = await validateDataContract(dataContract);
+    const result = await validateDataContract(rawDataContract);
 
     expectValidationError(result, InvalidIndexedPropertyConstraintError);
 
@@ -1620,7 +1618,7 @@ describe('validateDataContractFactory', () => {
   it('should return invalid result if indexed property have to big maxLength', async () => {
     rawDataContract.documents.indexedDocument.properties.firstName.maxLength = 2048;
 
-    const result = await validateDataContract(dataContract);
+    const result = await validateDataContract(rawDataContract);
 
     expectValidationError(result, InvalidIndexedPropertyConstraintError);
 
@@ -1632,7 +1630,7 @@ describe('validateDataContractFactory', () => {
   });
 
   it('should return valid result if Data Contract is valid', async () => {
-    const result = await validateDataContract(dataContract);
+    const result = await validateDataContract(rawDataContract);
 
     expect(result).to.be.an.instanceOf(ValidationResult);
     expect(result.isValid()).to.be.true();
